Add framework node to SDLStage and Lingo ovie and LingoSprite

+++ src/LingoEngine.IO.Data/DTO/LingoStageDTO.cs
@@ -3,7 +3,7 @@
 public class LingoStageDTO
 {
    
-    public int Width { get; set; } = 640;
-    public int Height { get; set; } = 480;
+    public float Width { get; set; } = 640;
+    public float Height { get; set; } = 480;
     public LingoColorDTO BackgroundColor { get; set; } = new LingoColorDTO { Name = "black" ,A = 255};
 }
